test: add example of how transform breaks backpressure
The 'transform' function seems to ignore backpressure from its
destination. As a result, when the source is writing data faster than
the destination can accept it, the transform's internal buffer grows
without limit.

To see this happen, navigate to packages/stream-transform and execute
'npm run test'. The new 'backpressure.js' sample should time out after
an hour. Instead, it runs out of memory pretty quick. If you comment
out the transform stream in that file, it works as expected.